requirements.txt
blinker==1.8.0
click==8.1.7
exceptiongroup==1.2.1
flask==3.0.3
importlib-metadata==7.1.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
kafka-python==2.0.2
MarkupSafe==2.1.5
numpy==1.26.4
opencv-contrib-python==4.9.0.80
packaging==24.0
pluggy==1.5.0
pytest==8.2.0
tomli==2.0.1
werkzeug==3.0.2
zipp==3.18.1
